Update to version 1.9

+++ source/projects/cv.jit.binedge/max.cv.jit.binedge.cpp
@@ -83,7 +83,7 @@ void *max_cv_jit_binedge_new(t_symbol *s, long argc, t_atom *argv)
 	if (x) {  //Create a new instance of Jitter object class
 		if ((o= (t_jit_object*)jit_object_new(gensym("cv_jit_binedge")))) { //Create a new binedge object
 			max_jit_mop_setup_simple(x,o,argc,argv);  //Matrix operator setup
-			max_jit_attr_args(x,argc,argv); //Arguments and attributes
+			max_jit_attr_args(x,(short)argc,argv); //Arguments and attributes
 		} else {
 			object_error((t_object*)x, "cv.jit.binedge: could not allocate object");
 			object_free((t_object *)x);
