+++ source/projects/cv.jit.blobs.centroids/max.cv.jit.blobs.centroids.cpp
@@ -80,7 +80,7 @@ void *max_cv_jit_blobs_centroids_new(t_symbol *s, long argc, t_atom *argv)
 	if ((x=(t_max_cv_jit_blobs_centroids *)max_jit_obex_new(max_cv_jit_blobs_centroids_class,gensym("cv_jit_blobs_centroids")))) {
 		if ((o= (t_jit_object*)jit_object_new(gensym("cv_jit_blobs_centroids")))) {
 			max_jit_mop_setup_simple(x,o,argc,argv);			
-			max_jit_attr_args(x,argc,argv);
+			max_jit_attr_args(x,(short)argc,argv);
 		} else {
 			object_error((t_object*)x, "cv.jit.blobs.centroids: could not allocate object");
 			object_free((t_object *)x);
